+++ pkg/config/config_darwin.go
@@ -1,9 +1,5 @@
 package config
 
-import (
-	"os"
-)
-
 const (
 	// OverrideContainersConfig holds the default config path overridden by the root user
 	OverrideContainersConfig = "/etc/" + _configPath
@@ -18,18 +14,6 @@ const (
 	// Mount type for mounting host dir
 	_typeBind = "bind"
 )
-
-// podman remote clients on darwin cannot use unshare.isRootless() to determine the configuration file locations.
-func customConfigFile() (string, error) {
-	if path, found := os.LookupEnv(containersConfEnv); found {
-		return path, nil
-	}
-	return rootlessConfigPath()
-}
-
-func ifRootlessConfigPath() (string, error) {
-	return rootlessConfigPath()
-}
 
 var defaultHelperBinariesDir = []string{
 	// Relative to the binary directory
