+++ pkg/config/config_linux.go
@@ -1,9 +1,6 @@
 package config
 
 import (
-	"os"
-
-	"github.com/containers/storage/pkg/unshare"
 	selinux "github.com/opencontainers/selinux/go-selinux"
 )
 
@@ -24,31 +21,6 @@ const (
 
 func selinuxEnabled() bool {
 	return selinux.GetEnabled()
-}
-
-func customConfigFile() (string, error) {
-	if path, found := os.LookupEnv(containersConfEnv); found {
-		return path, nil
-	}
-	if unshare.GetRootlessUID() > 0 {
-		path, err := rootlessConfigPath()
-		if err != nil {
-			return "", err
-		}
-		return path, nil
-	}
-	return OverrideContainersConfig, nil
-}
-
-func ifRootlessConfigPath() (string, error) {
-	if unshare.GetRootlessUID() > 0 {
-		path, err := rootlessConfigPath()
-		if err != nil {
-			return "", err
-		}
-		return path, nil
-	}
-	return "", nil
 }
 
 var defaultHelperBinariesDir = []string{
