+++ pkg/config/config_windows.go
@@ -17,15 +17,9 @@ const (
 	_typeBind = "bind"
 )
 
-// podman remote clients on windows cannot use unshare.isRootless() to determine the configuration file locations.
-func customConfigFile() (string, error) {
-	if path, found := os.LookupEnv(containersConfEnv); found {
-		return path, nil
-	}
-	return os.Getenv("APPDATA") + "\\containers\\containers.conf", nil
-}
-
-func ifRootlessConfigPath() (string, error) {
+// userConfigPath returns the path to the users local config that is
+// not shared with other users. It uses $APPDATA/containers...
+func userConfigPath() (string, error) {
 	return os.Getenv("APPDATA") + "\\containers\\containers.conf", nil
 }
