+++ pkg/config/containers.conf
@@ -10,7 +10,8 @@
 # locations in the following order:
 #  1. /usr/share/containers/containers.conf
 #  2. /etc/containers/containers.conf
-#  3. $HOME/.config/containers/containers.conf (Rootless containers ONLY)
+#  3. $XDG_CONFIG_HOME/containers/containers.conf or
+#     $HOME/.config/containers/containers.conf if $XDG_CONFIG_HOME is not set
 #  Items specified in the latter containers.conf, if they exist, override the
 # previous containers.conf settings, or the default settings.
 
@@ -348,7 +349,7 @@ default_sysctls = [
 #]
 
 # The firewall driver to be used by netavark.
-# The default is empty which means netavark will pick one accordingly. Current supported 
+# The default is empty which means netavark will pick one accordingly. Current supported
 # drivers are "iptables", "none" (no firewall rules will be created) and "firewalld" (firewalld is
 # experimental at the moment and not recommend outside of testing). In the future we are
 # planning to add support for a "nftables" driver.
